requirements.txt
streamlit==0.92.0
pandas==1.3.4
seaborn==0.11.2
matplotlib==3.5.1
Pillow==8.4.0
